requirements.txt
# Core Application Framework
streamlit>=1.28.0
openai>=1.3.0

# Data Processing and Analysis
pandas>=2.2.0
numpy>=1.24.0
plotly>=5.15.0

# API and Communication
asyncio-throttle>=1.0.0
python-dotenv>=1.0.0

# Data Validation and Management
pydantic>=2.5.0

# Standard Library Extensions (built-in, no installation needed)
# uuid
# datetime
# json
# typing

# Advanced Features
scikit-learn>=1.3.0
networkx>=3.2.0
seaborn>=0.13.0
matplotlib>=3.8.0

# Performance and Optimization
cachetools>=5.3.0
psutil>=5.9.0

# Additional Utilities (built-in, no installation needed)
# hashlib
# math
# random
# dataclasses
# enum 